src/hdp_system/evaluate.py
import pandas as pd
import joblib
import os

from sqlalchemy import create_engine
from dotenv import load_dotenv

from sklearn.model_selection import train_test_split
from sklearn.metrics import (
    accuracy_score, precision_score, recall_score,
    f1_score, roc_auc_score, confusion_matrix,
    classification_report
)

# ---------------------------------------
# Load DB + model
# ---------------------------------------
load_dotenv()
DATABASE_URL = os.getenv("DATABASE_URL")
engine = create_engine(DATABASE_URL)

df = pd.read_sql("SELECT * FROM features", engine)

numeric_features = [
    "age", "sex", "cp", "trestbps", "chol", "fbs",
    "restecg", "thalach", "exang", "oldpeak",
    "slope", "ca", "thal"
]
categorical_features = ["age_group"]

X = df[numeric_features + categorical_features]
y = df["target"].astype(int)

#
