requirements.txt
pandas
numpy
scikit-learn
xgboost
SQLAlchemy
psycopg2-binary
joblib
matplotlib
seaborn
shap
streamlit
python-dotenv
